feat(uds): complete validation docs and bounded test coverage

+++ tests/uds/test_uds_core.c
@@ -5,12 +5,13 @@
 #include "uds.h"
 
 #include <assert.h>
+#include <stddef.h>
 #include <stdbool.h>
 #include <stdint.h>
 #include <string.h>
 
-static UdsCallbackResult read_did(void *context, uint16_t did, uint8_t *data,
-                                  uint16_t *length, uint16_t capacity) {
+static UdsCallbackResult read_did(void *context, uint16_t did, uint8_t *data, uint16_t *length,
+                                  uint16_t capacity) {
     (void)context;
     if (did != 0xF190U) {
         return UDS_RESULT_OUT_OF_RANGE;
@@ -36,14 +37,13 @@ static UdsCallbackResult security_seed(void *context, uint8_t level, uint8_t *se
     return UDS_RESULT_OK;
 }
 
-static UdsCallbackResult security_key(void *context, uint8_t level,
-                                      const uint8_t *key, uint16_t length) {
+static UdsCallbackResult security_key(void *context, uint8_t level, const uint8_t *key,
+                                      uint16_t length) {
     (void)context;
     if ((level != 1U) || (length != 2U)) {
         return UDS_RESULT_INVALID_KEY;
     }
-    return ((key[0] == 0xCAU) && (key[1] == 0xFEU)) ? UDS_RESULT_OK
-                                                    : UDS_RESULT_INVALID_KEY;
+    return ((key[0] == 0xCAU) && (key[1] == 0xFEU)) ? UDS_RESULT_OK : UDS_RESULT_INVALID_KEY;
 }
 
 static UdsCallbackResult ecu_reset(void *context, uint8_t subfunction) {
@@ -52,21 +52,19 @@ static UdsCallbackResult ecu_reset(void *context, uint8_t subfunction) {
 }
 
 static UdsCallbackResult communication_control(void *context, uint8_t subfunction,
-                                                uint8_t communication_type) {
+                                               uint8_t communication_type) {
     (void)context;
-    return ((subfunction == 0U) && (communication_type == 0x01U))
-               ? UDS_RESULT_OK
-               : UDS_RESULT_OUT_OF_RANGE;
+    return ((subfunction == 0U) && (communication_type == 0x01U)) ? UDS_RESULT_OK
+                                                                  : UDS_RESULT_OUT_OF_RANGE;
 }
 
-static UdsCallbackResult routine_control(void *context, uint8_t subfunction,
-                                         uint16_t routine_id, const uint8_t *request,
-                                         uint16_t request_len, uint8_t *response,
-                                         uint16_t *response_len, uint16_t capacity) {
+static UdsCallbackResult routine_control(void *context, uint8_t subfunction, uint16_t routine_id,
+                                         const uint8_t *request, uint16_t request_len,
+                                         uint8_t *response, uint16_t *response_len,
+                                         uint16_t capacity) {
     (void)context;
     (void)request;
-    if ((subfunction != 1U) || (routine_id != 0x0203U) || (request_len != 1U) ||
-        (capacity < 1U)) {
+    if ((subfunction != 1U) || (routine_id != 0x0203U) || (request_len != 1U) || (capacity < 1U)) {
         return UDS_RESULT_OUT_OF_RANGE;
     }
     response[0] = 0xAAU;
@@ -74,8 +72,8 @@ static UdsCallbackResult routine_control(void *context, uint8_t subfunction,
     return UDS_RESULT_OK;
 }
 
-static UdsCallbackResult request_download(void *context, uint32_t address,
-                                           uint32_t length, uint16_t *max_block_length) {
+static UdsCallbackResult request_download(void *context, uint32_t address, uint32_t length,
+                                          uint16_t *max_block_length) {
     (void)context;
     if ((address != 0x08080000UL) || (length != 0x1000UL)) {
         return UDS_RESULT_OUT_OF_RANGE;
@@ -84,17 +82,16 @@ static UdsCallbackResult request_download(void *context, uint32_t address,
     return UDS_RESULT_OK;
 }
 
-static UdsCallbackResult transfer_data(void *context, uint8_t block,
-                                       const uint8_t *data, uint16_t length) {
+static UdsCallbackResult transfer_data(void *context, uint8_t block, const uint8_t *data,
+                                       uint16_t length) {
     (void)context;
     (void)data;
-    return ((block != 0U) && (length > 0U)) ? UDS_RESULT_OK
-                                            : UDS_RESULT_OUT_OF_RANGE;
+    return ((block != 0U) && (length > 0U)) ? UDS_RESULT_OK : UDS_RESULT_OUT_OF_RANGE;
 }
 
-static UdsCallbackResult transfer_exit(void *context, const uint8_t *request,
-                                       uint16_t request_len, uint8_t *response,
-                                       uint16_t *response_len, uint16_t capacity) {
+static UdsCallbackResult transfer_exit(void *context, const uint8_t *request, uint16_t request_len,
+                                       uint8_t *response, uint16_t *response_len,
+                                       uint16_t capacity) {
     (void)context;
     (void)request;
     if ((request_len != 0U) || (capacity < 1U)) {
@@ -116,8 +113,7 @@ static void test_isotp(void) {
 
     IsoTpCanFrame frame;
     uint8_t payload[] = {0x22U, 0xF1U, 0x90U};
-    assert(isotp_tx_start(&tx, payload, sizeof(payload), 0U, &frame) ==
-           ISOTP_TX_FRAME_READY);
+    assert(isotp_tx_start(&tx, payload, sizeof(payload), 0U, &frame) == ISOTP_TX_FRAME_READY);
     assert(frame.can_id == 0x7E8U && frame.dlc == 4U && frame.data[0] == 3U);
     IsoTpRxEvent event;
     assert(isotp_rx_feed(&rx, &frame, 0U, &event) == ISOTP_COMPLETE);
@@ -125,7 +121,7 @@ static void test_isotp(void) {
     assert(memcmp(event.payload, payload, sizeof(payload)) == 0);
 
     uint8_t long_payload[20];
-    for (uint8_t index = 0U; index < sizeof(long_payload); ++index) {
+    for (size_t index = 0U; index < sizeof(long_payload); ++index) {
         long_payload[index] = index;
     }
     assert(isotp_tx_start(&tx, long_payload, sizeof(long_payload), 0U, &frame) ==
@@ -146,8 +142,7 @@ static void test_isotp(void) {
         assert(status == ISOTP_TX_FRAME_READY);
         assert((frame.data[0] & 0x0FU) == expected_sequence);
         expected_sequence = (uint8_t)((expected_sequence + 1U) & 0x0FU);
-        assert(isotp_rx_feed(&rx, &frame, 1U, &event) ==
-               ((rx.active) ? ISOTP_OK : ISOTP_COMPLETE));
+        assert(isotp_rx_feed(&rx, &frame, 1U, &event) == ((rx.active) ? ISOTP_OK : ISOTP_COMPLETE));
         if (event.has_flow_control) {
             assert(isotp_tx_feed_flow_control(&tx, &event.flow_control, 1U) == ISOTP_OK);
         }
@@ -186,41 +181,40 @@ static void test_uds(void) {
     uint16_t response_len;
 
     uint8_t request[] = {0x10U, 0x03U};
-    assert(uds_server_handle(&server, request, sizeof(request), response,
-                             &response_len, sizeof(response), 0U) == UDS_RESULT_OK);
+    assert(uds_server_handle(&server, request, sizeof(request), response, &response_len,
+                             sizeof(response), 0U) == UDS_RESULT_OK);
     assert(response_len == 6U && response[0] == 0x50U && response[1] == 0x03U);
 
     uint8_t read_request[] = {0x22U, 0xF1U, 0x90U};
-    assert(uds_server_handle(&server, read_request, sizeof(read_request), response,
-                             &response_len, sizeof(response), 1U) == UDS_RESULT_OK);
+    assert(uds_server_handle(&server, read_request, sizeof(read_request), response, &response_len,
+                             sizeof(response), 1U) == UDS_RESULT_OK);
     assert(response_len == 10U && response[0] == 0x62U && response[1] == 0xF1U &&
            response[2] == 0x90U);
 
     uint8_t bad_did[] = {0x22U, 0x12U, 0x34U};
-    assert(uds_server_handle(&server, bad_did, sizeof(bad_did), response,
-                             &response_len, sizeof(response), 2U) == UDS_RESULT_OK);
+    assert(uds_server_handle(&server, bad_did, sizeof(bad_did), response, &response_len,
+                             sizeof(response), 2U) == UDS_RESULT_OK);
     assert(response_len == 3U && response[0] == 0x7FU && response[2] == 0x31U);
 
     uint8_t seed_request[] = {0x27U, 0x01U};
-    assert(uds_server_handle(&server, seed_request, sizeof(seed_request), response,
-                             &response_len, sizeof(response), 3U) == UDS_RESULT_OK);
+    assert(uds_server_handle(&server, seed_request, sizeof(seed_request), response, &response_len,
+                             sizeof(response), 3U) == UDS_RESULT_OK);
     assert(response_len == 4U && response[0] == 0x67U && response[2] == 0x12U);
     uint8_t key_request[] = {0x27U, 0x02U, 0xCAU, 0xFEU};
-    assert(uds_server_handle(&server, key_request, sizeof(key_request), response,
-                             &response_len, sizeof(response), 4U) == UDS_RESULT_OK);
+    assert(uds_server_handle(&server, key_request, sizeof(key_request), response, &response_len,
+                             sizeof(response), 4U) == UDS_RESULT_OK);
     assert(uds_server_security_level(&server) == 1U);
 
     uint8_t reset_request[] = {0x11U, 0x01U};
-    assert(uds_server_handle(&server, reset_request, sizeof(reset_request), response,
-                             &response_len, sizeof(response), 5U) == UDS_RESULT_OK);
+    assert(uds_server_handle(&server, reset_request, sizeof(reset_request), response, &response_len,
+                             sizeof(response), 5U) == UDS_RESULT_OK);
     assert(uds_server_reset_pending(&server));
     uds_server_clear_reset(&server);
     assert(!uds_server_reset_pending(&server));
 
     uint8_t communication_request[] = {0x28U, 0x00U, 0x01U};
-    assert(uds_server_handle(&server, communication_request,
-                             sizeof(communication_request), response, &response_len,
-                             sizeof(response), 6U) == UDS_RESULT_OK);
+    assert(uds_server_handle(&server, communication_request, sizeof(communication_request),
+                             response, &response_len, sizeof(response), 6U) == UDS_RESULT_OK);
     assert(response[0] == 0x68U);
 
     uint8_t routine_request[] = {0x31U, 0x01U, 0x02U, 0x03U, 0xAAU};
@@ -228,25 +222,23 @@ static void test_uds(void) {
                              &response_len, sizeof(response), 7U) == UDS_RESULT_OK);
     assert(response_len == 5U && response[0] == 0x71U && response[4] == 0xAAU);
 
-    uint8_t download_request[] = {0x34U, 0x44U, 0x08U, 0x08U, 0x00U, 0x00U,
-                                   0x00U, 0x00U, 0x10U, 0x00U};
-    assert(uds_server_handle(&server, download_request, sizeof(download_request),
-                             response, &response_len, sizeof(response), 8U) ==
-           UDS_RESULT_OK);
+    uint8_t download_request[] = {0x34U, 0x44U, 0x08U, 0x08U, 0x00U,
+                                  0x00U, 0x00U, 0x00U, 0x10U, 0x00U};
+    assert(uds_server_handle(&server, download_request, sizeof(download_request), response,
+                             &response_len, sizeof(response), 8U) == UDS_RESULT_OK);
     assert(response[0] == 0x74U && server.download_active);
     uint8_t transfer_request[] = {0x36U, 0x01U, 0xAAU, 0xBBU};
-    assert(uds_server_handle(&server, transfer_request, sizeof(transfer_request),
-                             response, &response_len, sizeof(response), 9U) ==
-           UDS_RESULT_OK);
+    assert(uds_server_handle(&server, transfer_request, sizeof(transfer_request), response,
+                             &response_len, sizeof(response), 9U) == UDS_RESULT_OK);
     assert(response[0] == 0x76U);
     uint8_t exit_request[] = {0x37U};
-    assert(uds_server_handle(&server, exit_request, sizeof(exit_request), response,
-                             &response_len, sizeof(response), 10U) == UDS_RESULT_OK);
+    assert(uds_server_handle(&server, exit_request, sizeof(exit_request), response, &response_len,
+                             sizeof(response), 10U) == UDS_RESULT_OK);
     assert(response_len == 2U && response[0] == 0x77U && response[1] == 0x55U);
 
     uint8_t unsupported[] = {0x99U};
-    assert(uds_server_handle(&server, unsupported, sizeof(unsupported), response,
-                             &response_len, sizeof(response), 11U) == UDS_RESULT_OK);
+    assert(uds_server_handle(&server, unsupported, sizeof(unsupported), response, &response_len,
+                             sizeof(response), 11U) == UDS_RESULT_OK);
     assert(response[0] == 0x7FU && response[1] == 0x99U && response[2] == 0x11U);
 }
